backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0
django-cors-headers>=4.4
psycopg[binary]>=3.2
python-dotenv>=1.0
Pillow>=10.0
gunicorn>=21.0
whitenoise>=6.6
dj-database-url>=2.1
